Overhaul parsing of macro definitions
Macro definitions may not be valid PIC code as they are resolved by
the lexer and only the code with the expanded macro is parsed. This
effectively prevents us from making anything useful from the macro
definition. So the complete macro body is parsed as an opaque stream
of characters.